+++ src/app/static/package.json
@@ -8,10 +8,10 @@
     "js": "webpack",
     "build": "gulp copy-templates && gulp styles && webpack",
     "test": "jest -c jest.config.js",
-    "integration-test": "jest --testPathIgnorePatterns adr -c jest.integration.config.js",
-    "adr-integration-test": "jest --testPathPattern adr -c jest.integration.adr.config.js",
-    "integration-test-docker": "jest --testPathIgnorePatterns adr -c jest.integration.docker.config.js",
-    "adr-integration-test-docker": "jest --testPathPattern adr -c jest.integration.docker.adr.config.js",
+    "integration-test": "jest --testPathIgnorePatterns adr-dataset -c jest.integration.config.js",
+    "adr-integration-test": "jest --testPathPattern adr-dataset -c jest.integration.adr.config.js",
+    "integration-test-docker": "jest --testPathIgnorePatterns adr-dataset -c jest.integration.docker.config.js",
+    "adr-integration-test-docker": "jest --testPathPattern adr-dataset -c jest.integration.docker.adr.config.js",
     "lint": "npx eslint src --ext .js,.jsx,.ts,.tsx,vue"
   },
   "author": "",
